+++ lib/src/remember_listenable.dart
@@ -6,13 +6,14 @@ import 'package:an_lifecycle_cancellable/an_lifecycle_cancellable.dart'
         ListenableCancellable,
         CancellableValueNotifier;
 import 'package:anlifecycle/anlifecycle.dart';
+import 'package:cancellable/cancellable.dart';
 import 'package:flutter/widgets.dart';
 import 'package:remember/src/remember.dart';
 
 extension RememberListenableExt on BuildContext {
   /// 快速生成一个可重用的 Listenable
   /// * 调用顺序、[T]和 [key] 确定是否为同一个对象 如果发生了变化则重新创建
-  /// * [value]， [factory]， [factory2] 确定如何初始化的创建一个 ValueNotifier 必须有一个不能为空 不作为更新key
+  /// * [value]， [factory]， [factory2]， [factory3] 确定如何初始化的创建一个 ValueNotifier 必须有一个不能为空 不作为更新key
   /// * [listen] 当前的 Context 自动监听生成的 ValueNotifier 只有首次有效 后续变化无效
   /// * [onCreate] 创建完成时的处理
   /// * [onDispose] 定义销毁时如何处理，晚于[context]的[dispose],**非常注意：不可使用[context]相关内容**
@@ -21,7 +22,8 @@ extension RememberListenableExt on BuildContext {
     T? value,
     T Function()? factory,
     T Function(Lifecycle)? factory2,
-    void Function(Lifecycle, T)? onCreate,
+    T Function(Lifecycle, Cancellable)? factory3,
+    void Function(T, Lifecycle, Cancellable)? onCreate,
     FutureOr<void> Function(T)? onDispose,
     bool listen = false,
     Object? key,
@@ -29,7 +31,8 @@ extension RememberListenableExt on BuildContext {
       remember<T>(
         factory: value == null ? factory : () => value,
         factory2: factory2,
-        onCreate: (l, v) {
+        factory3: factory3,
+        onCreate: (v, l, c) {
           if (listen && this is Element) {
             final rContext = WeakReference(this);
             v.addCListener(l.makeLiveCancellable(), () {
@@ -39,7 +42,7 @@ extension RememberListenableExt on BuildContext {
               }
             });
           }
-          onCreate?.call(l, v);
+          onCreate?.call(v, l, c);
         },
         onDispose: onDispose,
         key: key,
@@ -49,14 +52,15 @@ extension RememberListenableExt on BuildContext {
 extension RememberChangeNotifierExt on BuildContext {
   /// 快速生成一个可重用的 ChangeNotifier
   /// * 调用顺序、[T]和 [key] 确定是否为同一个对象 如果发生了变化则重新创建
-  /// * [factory]， [factory2] 确定如何初始化的创建一个 ValueNotifier 必须有一个不能为空 不作为更新key
+  /// * [factory]， [factory2]， [factory3] 确定如何初始化的创建一个 ValueNotifier 必须有一个不能为空 不作为更新key
   /// * [listen] 当前的 Context 自动监听生成的 ValueNotifier 只有首次有效 后续变化无效
   /// * [onCreate] 创建完成时的处理
   /// * [onDispose] 定义销毁时如何处理,已自动调用[dispose]，晚于[context]的[dispose],**非常注意：不可使用[context]相关内容**
   T rememberChangeNotifier<T extends ChangeNotifier>({
     T Function()? factory,
     T Function(Lifecycle)? factory2,
-    void Function(Lifecycle, T)? onCreate,
+    T Function(Lifecycle, Cancellable)? factory3,
+    void Function(T, Lifecycle, Cancellable)? onCreate,
     FutureOr<void> Function(T)? onDispose,
     bool listen = false,
     Object? key,
@@ -64,6 +68,7 @@ extension RememberChangeNotifierExt on BuildContext {
       rememberListenable(
           factory: factory,
           factory2: factory2,
+          factory3: factory3,
           onCreate: onCreate,
           onDispose: (d) {
             d.dispose();
@@ -76,26 +81,28 @@ extension RememberChangeNotifierExt on BuildContext {
 extension RememberValueNotifierExt on BuildContext {
   /// 快速生成一个可重用的 ValueNotifier
   /// * 调用顺序、[T]和 [key] 确定是否为同一个对象 如果发生了变化则重新创建
-  /// * [value]， [factory]， [factory2] 确定如何初始化的创建一个 ValueNotifier 必须有一个不能为空 不作为更新key
+  /// * [value]， [factory]， [factory2]， [factory3] 确定如何初始化的创建一个 ValueNotifier 必须有一个不能为空 不作为更新key
   /// * [listen] 当前的 Context 自动监听生成的 ValueNotifier 只有首次有效 后续变化无效
   ValueNotifier<T> rememberValueNotifier<T>({
     T? value,
     T Function()? factory,
     T Function(Lifecycle)? factory2,
-    void Function(Lifecycle, ValueNotifier<T>)? onCreate,
+    T Function(Lifecycle, Cancellable)? factory3,
+    void Function(ValueNotifier<T>, Lifecycle, Cancellable)? onCreate,
     FutureOr<void> Function(ValueNotifier<T>)? onDispose,
     bool listen = false,
     Object? key,
   }) =>
       rememberChangeNotifier<ValueNotifier<T>>(
-        factory2: (l) {
+        factory3: (l, c) {
           assert(
             value != null || factory != null || factory2 != null,
             'value and factory and factory2 cannot be null at the same time',
           );
           value ??= factory?.call();
           value ??= factory2?.call(l);
-          return CancellableValueNotifier(value as T, l.makeLiveCancellable());
+          value ??= factory3?.call(l, c);
+          return CancellableValueNotifier(value as T, c);
         },
         onCreate: onCreate,
         onDispose: onDispose,