+++ CHANGELOG.md
@@ -1,6 +1,15 @@
+## 1.3.0
+
+* bindLifecycle is marked as deprecated.
+* The newly added factory3 includes a cancellable, allowing the alive state to be verified.
+
+### Breaking Changes
+
+* Add a cancellable in the onCreate callback to verify the alive state.
+
 ## 1.2.0
 
-* Added bindLifecycle to bind lifecycle.
+* Added [bindLifecycle] to bind lifecycle.
 
 ## 1.1.0